Extract open diff baseline lookup helper
